fix: hmr invalidate error
